# doc/swagger.yaml
swagger: "2.0"
info:
  description: "API for GiveItAway vk mini app"
  version: "1.0.0"
  title: "Отдай даром"
basePath: "/api"
...
schemes:
  - "https"
  - "http"
paths:
...
  /post/find:
    get:
      tags:
        - "ad"
      summary: "Найти объявления"
      operationId: "findAds"
      produces:
        - "application/json"
      parameters:
        - name: "category"
          in: "query"
          type: "string"
          description: "Category where to search"
        - name: "author_id"
          in: "query"
          type: "string"
          description: "Ads of some user"
        - name: "region"
          in: "query"
          type: "string"
          description: "Region where to search"
        - name: "district"
          in: "query"
          type: "string"
          description: "District where to search"
        - name: "query"
          in: "query"
          type: "string"
          description: "Query"
        - name: "page"
          in: "query"
          type: "string"
          description: "page for pagination"
          required: true
        - name: "rows_per_page"
          in: "query"
          type: "string"
          description: "rows per page for pagination"
          required: true
        - name: "sort_by"
          in: "query"
          description: "sort by time (\"time\") or distance to current geoposition (\"geo\")"
          type: "string"
          # enum:
          #   - "time"
          #   - "geo"
          # default: "time"
        - name: radius
          in: "query"
          description: "radius for getting ads in radius from your place"
          type: "string"
        - name: lat
          description: "lattitude of user's geoposition"
          in: "query"
          type: "number"
        - name: long
          in: "query"
          type: "number"
          description: "longitude of user's geoposition"
      responses:
        200:
          description: "found smth"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/AdForUsers"
        400:
          description: "Invalid parameters"
        404:
          description: "Found nothing"
...
definitions:
  User:
    type: "object"
    properties:
      vk_id:
        type: "integer"
        format: "int64"
      carma:
        type: "integer"
      name:
        type: "string"
      surname:
        type: "string"
      photo_url:
        type: "string"
...
  AdForUsers:
    type: "object"
    properties:
      ad_id:
        type: "integer"
        format: "int64"
      author:
        $ref: '#/definitions/User'
      header:
        type: "string"
      region:
        type: "string"
      district:
        type: "string"
      ad_type:
        type: "string"
        enum:
          - "choice"
          - "random"
          - "auction"
      ls_enabled:
        type: "boolean"
      metro:
        type: string
      comments_enabled:
        type: "boolean"
      extra_enabled:
        type: "boolean"
      extra_field:
        type: "string"
      creation_date:
        type: "string"
        example: "01.02.2006 15:04"
      status:
        type: "string"
        enum: ["offer", "chosen", "closed"]
        default: "offer"
      pathes_to_photo:
        type: "array"
        items:
          $ref: "#/definitions/AdPhoto"
      category:
        type: "string"
      comments_count:
        type: integer
...
  AdPhoto:
    type: "object"
    properties:
      ad_photo_id:
        type: integer
      photo_url:
        type: string
        example: "photo1"
...